Remove YouCompleteMe from default Vim bundles.
On FreeBSD (at least), the package managed does a much better job of
updating updating ycm than $HOME/.vim/bundles ever could. Remove ycm
from the Vim bundles so that we stop trying to do things the Wrong Way
on the majority of systems that I use. For systems that need to do it
the slightly awkward way (macOS, I'm looking at you), I can always
manually checkout this repository but not put it into `.gitmodules`.